requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
cryptography==2.4.2
dj-database-url==0.5.0
Django==2.1.4
django-heroku==0.3.1
django-solo==1.1.3
djangorestframework==3.9.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
psycopg2==2.7.6.1
pycparser==2.19
python-telegram-bot==11.1.0
pytz==2018.7
scrypt==0.8.6
six==1.12.0
virtualenv==16.1.0
whitenoise==4.1.2
